+++ upcloud/load_balancer.go
@@ -34,8 +34,9 @@ const (
 	LoadBalancerConfiguredStatusStarted LoadBalancerConfiguredStatus = "started"
 	LoadBalancerConfiguredStatusStopped LoadBalancerConfiguredStatus = "stopped"
 
-	LoadBalancerCertificateBundleTypeManual  LoadBalancerCertificateBundleType = "manual"
-	LoadBalancerCertificateBundleTypeDynamic LoadBalancerCertificateBundleType = "dynamic"
+	LoadBalancerCertificateBundleTypeManual    LoadBalancerCertificateBundleType = "manual"
+	LoadBalancerCertificateBundleTypeDynamic   LoadBalancerCertificateBundleType = "dynamic"
+	LoadBalancerCertificateBundleTypeAuthority LoadBalancerCertificateBundleType = "authority"
 
 	LoadBalancerOperationalStatePending       LoadBalancerOperationalState = "pending"
 	LoadBalancerOperationalStateSetupAgent    LoadBalancerOperationalState = "setup-agent"
@@ -157,7 +158,7 @@ type LoadBalancerFrontendRule struct {
 	UpdatedAt time.Time             `json:"updated_at,omitempty"`
 }
 
-// LoadBalancerFrontendTLSConfig represents TLS configuration
+// LoadBalancerFrontendTLSConfig represents frontend TLS configuration
 type LoadBalancerFrontendTLSConfig struct {
 	Name                  string    `json:"name,omitempty"`
 	CertificateBundleUUID string    `json:"certificate_bundle_uuid,omitempty"`
@@ -167,8 +168,9 @@ type LoadBalancerFrontendTLSConfig struct {
 
 // LoadBalancerFrontendProperties represents frontend properties
 type LoadBalancerFrontendProperties struct {
-	TimeoutClient        int  `json:"timeout_client,omitempty"`
-	InboundProxyProtocol bool `json:"inbound_proxy_protocol"`
+	TimeoutClient        int   `json:"timeout_client,omitempty"`
+	InboundProxyProtocol bool  `json:"inbound_proxy_protocol"`
+	HTTP2Enabled         *bool `json:"http2_enabled,omitempty"`
 }
 
 // LoadBalancerBackend represents service backend
@@ -177,6 +179,7 @@ type LoadBalancerBackend struct {
 	Members    []LoadBalancerBackendMember    `json:"members"`
 	Resolver   string                         `json:"resolver,omitempty"`
 	Properties *LoadBalancerBackendProperties `json:"properties,omitempty"`
+	TLSConfigs []LoadBalancerBackendTLSConfig `json:"tls_configs,omitempty"`
 	CreatedAt  time.Time                      `json:"created_at,omitempty"`
 	UpdatedAt  time.Time                      `json:"updated_at,omitempty"`
 }
@@ -194,6 +197,14 @@ type LoadBalancerBackendMember struct {
 	UpdatedAt   time.Time                     `json:"updated_at,omitempty"`
 }
 
+// LoadBalancerBackendTLSConfig represents backend TLS configuration
+type LoadBalancerBackendTLSConfig struct {
+	Name                  string    `json:"name,omitempty"`
+	CertificateBundleUUID string    `json:"certificate_bundle_uuid,omitempty"`
+	CreatedAt             time.Time `json:"created_at,omitempty"`
+	UpdatedAt             time.Time `json:"updated_at,omitempty"`
+}
+
 // LoadBalancerBackendProperties represents backend properties
 type LoadBalancerBackendProperties struct {
 	TimeoutServer             int                              `json:"timeout_server,omitempty"`
@@ -207,6 +218,10 @@ type LoadBalancerBackendProperties struct {
 	HealthCheckExpectedStatus int                              `json:"health_check_expected_status,omitempty"`
 	StickySessionCookieName   string                           `json:"sticky_session_cookie_name,omitempty"`
 	OutboundProxyProtocol     LoadBalancerProxyProtocolVersion `json:"outbound_proxy_protocol,omitempty"`
+	TLSEnabled                *bool                            `json:"tls_enabled,omitempty"`
+	TLSVerify                 *bool                            `json:"tls_verify,omitempty"`
+	TLSUseSystemCA            *bool                            `json:"tls_use_system_ca,omitempty"`
+	HTTP2Enabled              *bool                            `json:"http2_enabled,omitempty"`
 }
 
 // LoadBalancerResolver represents domain name resolver
@@ -268,7 +283,7 @@ type LoadBalancerMatcherStringWithArgument struct {
 	IgnoreCase *bool                           `json:"ignore_case,omitempty"`
 }
 
-// LoadBalancerMatcherHost represents represents 'host' matcher
+// LoadBalancerMatcherHost represents 'host' matcher
 type LoadBalancerMatcherHost struct {
 	Value string `json:"value,omitempty"`
 }
@@ -387,7 +402,7 @@ type LoadBalancerNodeNetwork struct {
 	IPAddresses []LoadBalancerIPAddress `json:"ip_addresses,omitempty"`
 }
 
-// LoadBalancerNetwork represents network attached to loadbalancer
+// LoadBalancerFrontendNetwork represents network attached to loadbalancer
 type LoadBalancerFrontendNetwork struct {
 	Name string `json:"name,omitempty"`
 }